refactor: move getCityFromURL to utils/cityUrl to fix react-refresh lint
Amp-Thread-ID: https://ampcode.com/threads/T-019d3b7a-6333-743e-99bf-f8b4e00fa59d
